Moved files to directories & import fixes
