+++ app/src/main/java/com/ynovlyon/ymenu/presentation/dish_list/DishListItem.kt
@@ -1,6 +1,7 @@
 package com.ynovlyon.ymenu
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.CornerSize
 import androidx.compose.foundation.shape.RoundedCornerShape
