+++ app/src/main/java/com/ynovlyon/ymenu/presentation/navbar/NavigationViews.kt
@@ -3,10 +3,10 @@ package com.ynovlyon.ymenu
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import com.ynovlyon.ymenu.data.Dish
-import com.ynovlyon.ymenu.ui.theme.DishList
 import com.ynovlyon.ymenu.presentation.onboarding.Greeting
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.ynovlyon.ymenu.camera.Camera
+import com.ynovlyon.ymenu.presentation.dish_list.DishList
 
 @Composable
 fun Greeting(name: String) {
