+++ app/src/main/java/com/ynovlyon/ymenu/ui/theme/Color.kt
@@ -7,5 +7,4 @@ val Purple500 = Color(0xFF6200EE)
 val Purple700 = Color(0xFF3700B3)
 val Teal200 = Color(0xFF03DAC5)
 val graySurface = Color(0xFF202124)
-val Teal200 = Color(0xFFEDA751)
 val Orange200 = Color (0xFFFDAF5E)
